+++ metadata.rb
@@ -14,4 +14,4 @@ recipe           'openstack-identity::registration', 'Adds user, tenant, role an
   supports os
 end
 
-depends          'openstack-common', '~> 10.0'
+depends          'openstack-common', '~> 10.1'